Feature: display basic weather info
When the user searchs a city, it will display an icon, along with a
text that describes the current condition and the temperature (set to
celsius for the moment).

In the form, moved the addFunctionality function to dom module to
call inside this function another one that uses the data to display it
on the page.

Added a new module to help separate the code and a json file to use
to get the proper icon.