Feature: let the user switch temperature units
To make this feature work, added a button that's placed before the form.
The form now has an input "hidden" element that contains the temperature
unit that will be used to display the information when it's required.

Also changed the previous line where it shows this temperature to a
function that gets the proper temperature.

At the moment, if the user clicks the button when it searched a city,
the temperature unit don't change.