Refactor: code in dom.js, form.js
Changed some lines to prevent repetition.

Changed the getCurrentTemperature function to receive an object with two
temperatures to make it reusable.

Added listener to button that switchs temperatures to update them when
clicked. The function is stored inside the addFunctionality function,
where the weather data is accessible.

Also move some code into new functions to reduce a function size.